work stealing protocol
